dashplayer: CEA support in dashplayer using mediacodec, other fixes
1) CEA support in dashplayer using mediacodec
2) NULL check and other Code Cleanup in Dashplayer
3) handle flushcompleted state check

+++ dashplayer/DashPlayerStats.h
@@ -35,7 +35,6 @@
 #include <utils/threads.h>
 #include <utils/Log.h>
 
-
 namespace android {
 
 class DashPlayerStats : public RefBase {
@@ -73,11 +72,11 @@ class DashPlayerStats : public RefBase {
     int64_t mNumVideoFramesDecoded;
     int64_t mNumVideoFramesDropped;
     int64_t mConsecutiveFramesDropped;
-    uint32_t mCatchupTimeStart;
+    int64_t mCatchupTimeStart;
     uint32_t mNumTimesSyncLoss;
-    uint32_t mMaxEarlyDelta;
-    uint32_t mMaxLateDelta;
-    uint32_t mMaxTimeSyncLoss;
+    int64_t mMaxEarlyDelta;
+    int64_t mMaxLateDelta;
+    int64_t mMaxTimeSyncLoss;
     uint64_t mTotalFrames;
     int64_t mFirstFrameLatencyStartUs;
     int64_t mLastFrame;
